+++ GigasServer/src/org/gigas/core/server/thread/ProtoBufBasedMessageSenderThread.java
@@ -57,8 +57,8 @@ public class ProtoBufBasedMessageSenderThread extends Thread implements IThread 
 							buf.writeBytes(byteArray);
 							ByteBuf result = channel.alloc().directBuffer();
 							result.writeBytes(secirityBytes);
-							result.writeInt(Long.SIZE / Byte.SIZE + buf.readableBytes());
-							result.writeLong(poll.getId());
+							result.writeInt(Integer.SIZE / Byte.SIZE + buf.readableBytes());
+							result.writeInt(poll.getId());
 							result.writeBytes(buf);
 							channel.writeAndFlush(result);
 							buf.release();
